move isc_socket_connect() calls into dispatch
dns_dispatch_connect() connects a dispatch socket (for TCP)
or a dispatch entry socket (for UDP). This is the next step
in moving all uses of the isc_socket code into the dispatch
module.

This API is temporary; it needs to be cleaned up further so
that it can be called the same way for both TCP and UDP.