add query to queries list before connecting, not after
it's possible for fctx_doshutdown() to run before a TCP connection
has completed. if the query is not on the queries list, then it is
not canceled, but the adbaddrinfo is freed. when tcp_connected() runs
later, the query is in an inconstent state. to fix this, we add the
query to queries before running dns_dispatch_connect(), instead of in
the connect callback.

incidental fix, noticed while working on this: new_adbaddrinfo() can't
return NULL anymore, so the code can be simplified.