send shutdown event if canceled with an active dispentry
observed in the 'legacy' system test that tcp_recv() can be
called with ISC_R_CANCELED while actively reading, for example
because of the netmgr shutting down. if there's more than one
dispentry waiting for this connection we end up with dangling
references. we clean up by calling the response callbacks for
all the listening dispentries with ISC_R_SHUTTINGDOWN.